+++ PSProfile/Public/Plugins/Add-PSProfilePlugin.ps1
@@ -41,7 +41,11 @@ function Add-PSProfilePlugin {
             if ($PSBoundParameters.ContainsKey('ArgumentList')) {
                 $plugin['ArgumentList'] = $ArgumentList
             }
-            $temp = @(($Global:PSProfile.Plugins | Where-Object {$_.Name -ne $pName}),$plugin)
+            $temp = @()
+            $Global:PSProfile.Plugins | Where-Object {$_.Name -ne $pName} | ForEach-Object {
+                $temp += $_
+            }
+            $temp += $plugin
             $Global:PSProfile.Plugins = $temp
         }
         if ($Save) {